Move function out of module index
And in to dedicated file.
