+++ src/harpstrata/Pitch/index.ts
@@ -1,7 +1,5 @@
-import type { HalfstepIndexMatrix } from '../Apparatus'
-
-import type { PitchIds, PitchMatrix, Pitch } from './types'
-import { C, Db, D, Eb, E, F, Gb, G, Ab, A, Bb, B, ORDERED_PITCHES } from './constants'
+import type { PitchIds, Pitch } from './types'
+import { C, Db, D, Eb, E, F, Gb, G, Ab, A, Bb, B } from './constants'
 
 const pitchMap = new Map()
 pitchMap.set(C.id, C)
@@ -20,20 +18,7 @@ pitchMap.set(B.id, B)
 export const getActivePitchIds = (): PitchIds[] => Array.from(pitchMap.keys())
 export const getPitch = (pitchId: PitchIds): Pitch => pitchMap.get(pitchId)
 
-export const getPitchMatrix = (halfstepIndexMatrix: HalfstepIndexMatrix, keyPitchId: PitchIds): PitchMatrix => {
-  const orderedPitchIds = ORDERED_PITCHES.map(pitch => pitch.id)
-  const pitchIndex = orderedPitchIds.indexOf(keyPitchId)
-  const arrayHead = [ ...ORDERED_PITCHES.slice(pitchIndex) ]
-  const arrayTail = [ ...ORDERED_PITCHES.slice(0, (pitchIndex)) ]
-  const alignedPitchIds = [ ...arrayHead, ...arrayTail ]
-
-  return halfstepIndexMatrix.map((halfstepIndexRow) => {
-    return halfstepIndexRow.map((halfstepIndex) => {
-      if (halfstepIndex === undefined) return undefined
-      return alignedPitchIds[halfstepIndex % 12]
-    })
-  })
-}
+export { getPitchMatrix } from './getPitchMatrix'
 
 export type { Pitch, PitchRow, PitchMatrix } from './types'
 export { PitchIds } from './types'
